[add]自分たちのScriptをprojectファルダの下に置いた
