+++ Assets/Project/Script/CommunicationManager.cs
