+++ Assets/Project/VrScenes/PlayButton_Stop.png.meta
